# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/users/me:
    get:
      operationId: UsersController_getMe
      summary: Get current authenticated user
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - users
      security:
        - access-token: []
    patch:
      operationId: UsersController_updateMe
      summary: Update current authenticated user profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users/me/stats:
    get:
      operationId: UsersController_getMyStats
      summary: Get gamification stats for the current user
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users/me/quests/history:
    get:
      operationId: UsersController_getMyQuestHistory
      summary: List quest history for the current user
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Free-text search query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            default: COMPLETED
            enum:
              - STARTED
              - COMPLETED
              - ABANDONED
              - EXPIRED
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestHistoryItemDto"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users/me/avatar:
    post:
      operationId: UsersController_uploadAvatar
      summary: Upload (or replace) the current user avatar
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: ""
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users:
    get:
      operationId: UsersController_list
      summary: List users (paginated)
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Free-text search query
          schema:
            type: string
        - name: sort
          required: false
          in: query
          description: Stable ordering for discovery; ties break by user id ascending. Defaults to CREATED_DESC.
          schema:
            $ref: "#/components/schemas/UserListSort"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserDto"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users/{id}:
    get:
      operationId: UsersController_getById
      summary: Get user by id
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users/{id}/stats:
    get:
      operationId: UsersController_getStats
      summary: Get public gamification stats for a user
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - users
      security:
        - access-token: []
  /api/v1/users/{id}/quests/history:
    get:
      operationId: UsersController_getUserQuestHistory
      summary: List quest history for a user (public stats)
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Free-text search query
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            default: COMPLETED
            enum:
              - STARTED
              - COMPLETED
              - ABANDONED
              - EXPIRED
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestHistoryItemDto"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
      tags:
        - users
      security:
        - access-token: []
...
info:
  title: SideQuest API
  description: Gamified exploration platform - REST API
  version: "1"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the access token returned by /auth/login
  schemas:
...
    UpdateUserDto:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: number
        limit:
          type: number
        total:
          type: number
        totalPages:
          type: number
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
    QuestHistoryQuestSnapshotDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        coverImageUrl:
          type: object
      required:
        - id
        - slug
        - title
    QuestHistoryItemDto:
      type: object
      properties:
        id:
          type: string
        questId:
          type: string
        quest:
          $ref: "#/components/schemas/QuestHistoryQuestSnapshotDto"
        status:
          type: string
          enum:
            - STARTED
            - COMPLETED
            - ABANDONED
            - EXPIRED
        startedAt:
          format: date-time
          type: string
        completedAt:
          type: object
        xpEarned:
          type: number
        durationMinutes:
          type: object
      required:
        - id
        - questId
        - quest
        - status
        - startedAt
        - xpEarned
    UserDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        displayName:
          type: object
        avatarUrl:
          type: object
        bio:
          type: object
        role:
          type: string
          enum:
            - USER
            - MODERATOR
            - ADMIN
            - BUSINESS_OWNER
        status:
          type: string
          enum:
            - ACTIVE
            - SUSPENDED
            - DELETED
        xp:
          type: number
        level:
          type: number
        questsDone:
          type: number
        title:
          type: object
          nullable: true
          description: Gamified rank from level and quest count (capped by the lower tier), unless `titleOverride` is set
        streakDays:
          type: number
        inviteLink:
          type: string
          description: Public registration URL using username ref (separate from single-use email invite tokens)
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - username
        - role
        - status
        - xp
        - level
        - questsDone
        - title
        - streakDays
        - inviteLink
        - createdAt
        - updatedAt
    UserListSort:
      type: string
      enum:
        - XP_DESC
        - XP_ASC
        - LEVEL_DESC
        - CREATED_DESC
        - USERNAME_ASC
...